almacenes agregados al 100%

+++ database/migrations/2020_02_07_121755_create_c_facts_table.php
@@ -21,7 +21,7 @@ class CreateCFactsTable extends Migration
             $table->string('ruc_cliente');
             $table->string('dir_cliente');
             $table->string('razon');
-            $table->integer('id_user');
+            $table->unsignedBigInteger('id_user');
             $table->string('fecha');
             $table->string('tipo_venta');
             $table->string('serie');
